+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/MinApiTests/IntegrationTests/AuthorizedEndpointsIntegrationTests.cs
@@ -4,45 +4,44 @@ using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
 using MinApiTests.IntegrationTests.Helpers;
 
-namespace MinApiTests.IntegrationTests
+namespace MinApiTests.IntegrationTests;
+
+public class AuthorizedEndpointsIntegrationTests : IClassFixture<TestWebApplicationFactory<Program>>
 {
-    public class AuthorizedEndpointsIntegrationTests : IClassFixture<TestWebApplicationFactory<Program>>
+    private readonly TestWebApplicationFactory<Program> _factory;
+
+    public AuthorizedEndpointsIntegrationTests(TestWebApplicationFactory<Program> factory)
     {
-        private readonly TestWebApplicationFactory<Program> _factory;
+        _factory = factory;
+    }
 
-        public AuthorizedEndpointsIntegrationTests(TestWebApplicationFactory<Program> factory)
-        {
-            _factory = factory;
-        }
+    public static IEnumerable<object[]> AdminFlags => new List<object[]>
+    {
+        new object[] { "false", HttpStatusCode.Forbidden },
+        new object[] { "true", HttpStatusCode.OK }
+    };
 
-        public static IEnumerable<object[]> AdminFlags => new List<object[]>
+    [Theory]
+    [MemberData(nameof(AdminFlags))]
+    public async Task GetAdminEndpointIsReturnedForAnAuthorizedRequest(string isAdmin, HttpStatusCode code)
+    {
+        // Arrange
+        var client = _factory.WithWebHostBuilder(builder =>
         {
-            new object[] { "false", HttpStatusCode.Forbidden },
-            new object[] { "true", HttpStatusCode.OK }
-        };
-
-        [Theory]
-        [MemberData(nameof(AdminFlags))]
-        public async Task GetAdminEndpointIsReturnedForAnAuthorizedRequest(string isAdmin, HttpStatusCode code)
-        {
-            // Arrange
-            var client = _factory.WithWebHostBuilder(builder =>
+            builder.ConfigureTestServices(services =>
             {
-                builder.ConfigureTestServices(services =>
-                {
-                    services.AddAuthentication("Test")
-                        .AddScheme<TestAuthenticationSchemeOptions, TestAuthHandler>("Test",
-                            options => options.IsAdmin = isAdmin);
-                });
-            }).CreateClient();
+                services.AddAuthentication("Test")
+                    .AddScheme<TestAuthenticationSchemeOptions, TestAuthHandler>("Test",
+                        options => options.IsAdmin = isAdmin);
+            });
+        }).CreateClient();
 
-            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Test");
+        client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Test");
 
-            //Act
-            var response = await client.GetAsync("/admin");
+        //Act
+        var response = await client.GetAsync("/admin");
 
-            // Assert
-            Assert.Equal(code, response.StatusCode);
-        }
+        // Assert
+        Assert.Equal(code, response.StatusCode);
     }
 }
