+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/MinApiTests/IntegrationTests/Helpers/TestEmailService.cs
@@ -1,13 +1,12 @@
-using todo_group.Services;
+using WebMinRouteGroup.Services;
 
-namespace MinApiTests.IntegrationTests.Helpers
+namespace MinApiTests.IntegrationTests.Helpers;
+
+public class TestEmailService : IEmailService
 {
-    public class TestEmailService : IEmailService
+    public Task Send(string emailAddress, string body)
     {
-        public Task Send(string emaidAddress, string body)
-        {
-            // You don't want to send real email when running integration tests
-            return Task.CompletedTask;
-        }
+        // You don't want to send real email when running integration tests
+        return Task.CompletedTask;
     }
 }
