+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/MinApiTests/IntegrationTests/Helpers/TestWebApplicationFactory.cs
@@ -2,32 +2,31 @@ using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using todo_group.Data;
+using WebMinRouteGroup.Data;
 
-namespace MinApiTests.IntegrationTests.Helpers
+namespace MinApiTests.IntegrationTests.Helpers;
+
+public class TestWebApplicationFactory<TProgram>
+    : WebApplicationFactory<TProgram> where TProgram : class
 {
-    public class TestWebApplicationFactory<TProgram>
-        : WebApplicationFactory<TProgram> where TProgram : class
+    protected override IHost CreateHost(IHostBuilder builder)
     {
-        protected override IHost CreateHost(IHostBuilder builder)
+        builder.ConfigureServices(services =>
         {
-            builder.ConfigureServices(services =>
+            var descriptor = services.SingleOrDefault(d => d.ServiceType == typeof(DbContextOptions<TodoGroupDbContext>));
+
+            if (descriptor != null)
             {
-                var descriptor = services.SingleOrDefault(d => d.ServiceType == typeof(DbContextOptions<TodoGroupDbContext>));
+                services.Remove(descriptor);
+            }
 
-                if (descriptor != null)
-                {
-                    services.Remove(descriptor);
-                }
-
-                services.AddDbContext<TodoGroupDbContext>(options =>
-                {
-                    var path = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-                    options.UseSqlite($"Data Source={Path.Join(path, "todo_group_tests.db")}");
-                });
+            services.AddDbContext<TodoGroupDbContext>(options =>
+            {
+                var path = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+                options.UseSqlite($"Data Source={Path.Join(path, "WebMinRouteGroup_tests.db")}");
             });
+        });
 
-            return base.CreateHost(builder);
-        }
+        return base.CreateHost(builder);
     }
 }
