+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/MinApiTests/UnitTests/Helpers/MockDb.cs
@@ -1,13 +1,15 @@
 using Microsoft.EntityFrameworkCore;
-using todo_group.Data;
+using WebMinRouteGroup.Data;
+
+namespace MinApiTests.UnitTests.Helpers;
 
 public class MockDb : IDbContextFactory<TodoGroupDbContext>
 {
     public TodoGroupDbContext CreateDbContext()
     {
         var options = new DbContextOptionsBuilder<TodoGroupDbContext>()
-                .UseInMemoryDatabase($"InMemoryTestDb-{DateTime.Now.ToFileTimeUtc()}")
-                .Options;
+            .UseInMemoryDatabase($"InMemoryTestDb-{DateTime.Now.ToFileTimeUtc()}")
+            .Options;
 
         return new TodoGroupDbContext(options);
     }
