+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/MinApiTests/UnitTests/TodoMoqTests.cs
@@ -1,8 +1,8 @@
 using Microsoft.AspNetCore.Http.HttpResults;
 using Moq;
-using todo_group;
-using todo_group.Data;
-using todo_group.Services;
+using WebMinRouteGroup;
+using WebMinRouteGroup.Data;
+using WebMinRouteGroup.Services;
 
 namespace MinApiTests.UnitTests;
