+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/WebMinRouteGroup/Data/TodoDto.cs
@@ -1,4 +1,4 @@
-namespace todo_group.Data;
+namespace WebMinRouteGroup.Data;
 
 public class TodoDto
 {
