+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/WebMinRouteGroup/Services/EmailService.cs
@@ -1,0 +1,10 @@
+namespace WebMinRouteGroup.Services;
+
+public class EmailService : IEmailService
+{
+    public Task Send(string emailAddress, string body)
+    {
+        // Code for sending mails to a configured host
+        return Task.CompletedTask;
+    }
+}
