+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/WebMinRouteGroup/Services/IEmailService.cs
@@ -1,0 +1,6 @@
+namespace WebMinRouteGroup.Services;
+
+public interface IEmailService
+{
+    Task Send(string emailAddress, string body);
+}
