+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/WebMinRouteGroup/Services/ITodoService.cs
@@ -1,0 +1,18 @@
+using WebMinRouteGroup.Data;
+
+namespace WebMinRouteGroup.Services;
+
+public interface ITodoService
+{
+    Task<List<Todo>> GetAll();
+
+    Task<List<Todo>> GetIncompleteTodos();
+
+    ValueTask<Todo?> Find(int id);
+
+    Task Add(Todo todo);
+
+    Task Update(Todo todo);
+
+    Task Remove(Todo todo);
+}
