+++ aspnetcore/fundamentals/minimal-apis/7.0-samples/WebMinRouteGroup/appsettings.json
