--- aspnetcore/fundamentals/minimal-apis/7.0-samples/todo-group/Services/EmailService.cs
@@ -1,11 +1,0 @@
-namespace todo_group.Services
-{
-    public class EmailService : IEmailService
-    {
-        public Task Send(string emaidAddress, string body)
-        {
-            // Code for sending mails to a configured host
-            return Task.CompletedTask;
-        }
-    }
-}
