--- aspnetcore/fundamentals/minimal-apis/7.0-samples/todo-group/Services/IEmailService.cs
@@ -1,7 +1,0 @@
-namespace todo_group.Services
-{
-    public interface IEmailService
-    {
-        Task Send(string emaidAddress, string body);
-    }
-}
