--- aspnetcore/fundamentals/minimal-apis/7.0-samples/todo-group/Utilities.cs
@@ -1,24 +1,0 @@
-using todo_group.Data;
-
-namespace todo_group
-{
-    public static class Utilities
-    {
-        public static Dictionary<string, string[]> IsValid(TodoDto td)
-        {
-            Dictionary<string, string[]> errors = new();
-
-            if (string.IsNullOrEmpty(td.Title))
-            {
-                errors.TryAdd("todo.name.errors", new[] { "Name is empty" });
-            }
-
-            if (td.Title!.Length < 3)
-            {
-                errors.TryAdd("todo.name.errors", new[] { "Name length < 3" });
-            }
-
-            return errors;
-        }
-    }
-}
